+++ test/run-command.test.ts
@@ -3,7 +3,7 @@ import {join} from 'node:path'
 
 import {runCommand} from '../src'
 
-describe('command', () => {
+describe('runCommand', () => {
   // eslint-disable-next-line unicorn/prefer-module
   const root = join(__dirname, 'fixtures/multi')
 
@@ -23,6 +23,19 @@ describe('command', () => {
     const {result, stdout} = await runCommand<{name: string}>(['foo bar', '--name=foo'], {root})
     expect(stdout).to.equal('hello foo!\n')
     expect(result?.name).to.equal('foo')
+  })
+
+  it('should handle single string', async () => {
+    const {result, stdout} = await runCommand<{name: string}>('foo:bar --name=foo', {root})
+    expect(stdout).to.equal('hello foo!\n')
+    expect(result?.name).to.equal('foo')
+  })
+
+  it('should handle expected exit codes', async () => {
+    const {error, stdout} = await runCommand(['exit', '--code=101'], {root})
+    expect(stdout).to.equal('exiting with code 101\n')
+    expect(error?.message).to.equal('EEXIT: 101')
+    expect(error?.oclif?.exit).to.equal(101)
   })
 })
