+++ test/run-hook.test.ts
@@ -6,7 +6,7 @@ import {runHook} from '../src'
 // eslint-disable-next-line unicorn/prefer-module
 const root = join(__dirname, 'fixtures/multi')
 
-describe('hooks', () => {
+describe('runHook', () => {
   it('should run a hook', async () => {
     const {stdout} = await runHook('foo', {argv: ['arg']}, {root})
     expect(stdout).to.equal('foo hook args: arg\n')
